services/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
PyPDF2==3.0.1
python-docx==1.1.0
transformers>=4.21.0
torch>=1.12.0
spacy>=3.4.0
accelerate>=0.20.0

# Database dependencies
prisma==0.15.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# RegTech pipeline dependencies
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
numpy>=1.24.0

# LLM dependencies
openai>=1.0.0
python-dotenv>=1.0.0
